make check_url follow 301 redirect
Website for some gems (SequenceServer, for example) employ a 301 redirect to
direct the user to the actual home page.  Failing to follow redirects, when
checking validity of a URL (in the `check_url` method), causes Biogems.info to
use the rubydoc.info page for the gem as the project home page, which may not
always be desirable.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>